pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "SiliconeCenter",email = "infra@siliconecenter.com.br"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "match (>=0.3.2,<0.4.0)",
    "case (>=1.5.3,<2.0.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "openai (>=2.26.0,<3.0.0)",
    "fastapi-pagination (>=0.15.10,<0.16.0)"
]

[tool.poetry]
packages = [{include = "app", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pre-commit (>=4.5.1,<5.0.0)",
    "ruff (>=0.15.4,<0.16.0)",
    "black (>=26.1.0,<27.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "isort (>=8.0.1,<9.0.0)"
]
